requirements.txt
uvicorn
fastapi
pydantic014477